requirement.txt
streamlit==1.38.0 
langchain-groq==0.1.9 
sentence-transformers==3.1.1 
faiss-cpu==1.8.0 
PyPDF2==3.0.1 
pandas==2.2.3 
python-dotenv==1.0.1 
numpy==1.26.4 
torch==2.3.1